@@ -25,46 +25,49 @@ dependencies:
     sdk: flutter
 
   # https://pub.dev/packages/dio
-  dio: ^3.0.10
+  dio: ^4.0.0
   # https://pub.dev/packages/provider
-  provider: ^4.3.2+2
+  provider: ^5.0.0
   # https://pub.dev/packages/shared_preferences
-  shared_preferences: ^0.5.12+4
+  shared_preferences: ^2.0.6
   # https://pub.dev/packages/intl
-  intl: ^0.16.1
+  intl: ^0.17.0
   # https://pub.dev/packages/duration
-  duration: ^2.0.15
+  duration: ^3.0.6
   # https://pub.dev/packages/google_sign_in
-  google_sign_in: ^4.5.6
+  google_sign_in: ^5.0.4
   # https://pub.dev/packages/web_socket_channel
-  web_socket_channel: ^1.1.0
+  web_socket_channel: ^2.1.0
   # https://pub.dev/packages/crypto
-  crypto: ^2.1.5
+  crypto: ^3.0.1
   # https://pub.dev/packages/firebase_messaging
-  firebase_messaging: ^7.0.3
+  firebase_messaging: ^10.0.2
   # https://pub.dev/packages/firebase_analytics
-  firebase_analytics: ^6.3.0
+  firebase_analytics: ^8.1.2
   # https://pub.dev/packages/firebase_performance
-  firebase_performance: ^0.4.3
+  firebase_performance: ^0.7.0+5
   # https://pub.dev/packages/overlay_support
-  overlay_support: ^1.0.5
+  overlay_support: ^1.2.1
   # https://pub.dev/packages/image_picker
-  image_picker: ^0.6.7+17
+  image_picker: ^0.8.1+1
   # https://pub.dev/packages/sentry
-  sentry_flutter: ^4.0.1
+  sentry_flutter: ^5.1.0
   # https://pub.dev/packages/stomp_dart_client
-  stomp_dart_client: ^0.3.8
+  stomp_dart_client: ^0.4.3
   # https://pub.dev/packages/path
-  path: ^1.7.0
+  path: ^1.8.0
   # https://pub.dev/packages/mime
-  mime: ^0.9.7
+  mime: ^1.0.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  rename: 1.2.0
-  flutter_launcher_icons: 0.8.0
-  flutter_launcher_name: 0.0.1
+  rename: 1.3.1
+  flutter_launcher_icons: 0.9.0
+  flutter_launcher_name:
+    git:
+      url: https://github.com/ride4sun/flutter_launcher_name
+      ref: "0.9.1"
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
