@@ -21,7 +21,7 @@ class NotificationHandler extends StatefulWidget {
 }
 
 class _NotificationHandlerState extends State<NotificationHandler> {
-  final FirebaseMessaging _fcm = FirebaseMessaging();
+  final FirebaseMessaging _fcm = FirebaseMessaging.instance;
   StreamSubscription iosSubscription;
 
   VoidCallback tokenListener;
@@ -29,43 +29,33 @@ class _NotificationHandlerState extends State<NotificationHandler> {
   @override
   void initState() {
     super.initState();
-    if (Platform.isIOS) {
-      iosSubscription = _fcm.onIosSettingsRegistered.listen((data) {
-        // save the token  OR subscribe to a topic here
-      });
 
-      _fcm.requestNotificationPermissions(IosNotificationSettings());
+    if (Platform.isIOS) {
+      // todo we need a permission after that listen.
+      _fcm.requestPermission();
     }
 
-    _fcm.configure(
-      onMessage: (Map<String, dynamic> message) async {
-        print("onMessage: $message");
-        showOverlayNotification((context) {
-          return GestureDetector(
-            onVerticalDragUpdate: (details) {
-              if (details.delta.dy > -1.4) return;
-              // on swipe up
-              OverlaySupportEntry.of(context).dismiss();
-            },
-            child: Card(
-              margin: const EdgeInsets.symmetric(horizontal: 4),
-              child: SafeArea(
-                child: ListTile(
-                  title: Text(message['notification']['title']),
-                  subtitle: Text(message['notification']['body']),
-                ),
+    FirebaseMessaging.onMessage.listen((event) {
+      print("Firebase Messaging onMessage: $event");
+      showOverlayNotification((context) {
+        return GestureDetector(
+          onVerticalDragUpdate: (details) {
+            if (details.delta.dy > -1.4) return;
+            // on swipe up
+            OverlaySupportEntry.of(context).dismiss();
+          },
+          child: Card(
+            margin: const EdgeInsets.symmetric(horizontal: 4),
+            child: SafeArea(
+              child: ListTile(
+                title: Text(event.data['notification']['title']),
+                subtitle: Text(event.data['notification']['body']),
               ),
             ),
-          );
-        }, duration: Duration(seconds: 10));
-      },
-      onLaunch: (Map<String, dynamic> message) async {
-        print("onLaunch: $message");
-      },
-      onResume: (Map<String, dynamic> message) async {
-        print("onResume: $message");
-      },
-    );
+          ),
+        );
+      }, duration: Duration(seconds: 10));
+    });
 
     var authenticationState = context.read<AuthenticationState>();
     tokenListener = () {
